+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file based on the [Keep a Changelog](http://keepachangelog.com/) Standard. This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased](https://github.com/kununu/elasticsearch/compare/v2.1.2...master)
+## [Unreleased](https://github.com/kununu/elasticsearch/compare/v2.2...master)
+### Backward Compatibility Breaks
+### Bugfixes
+### Added
+### Improvements
+### Deprecated
+
+## [2.2](https://github.com/kununu/elasticsearch/compare/v2.1.2...v2.2)
 ### Backward Compatibility Breaks
 ### Bugfixes
 ### Added